requirements.txt
# Unilever Procurement GPT POC - Dependencies
# Python 3.11+

# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.32.1
pydantic==2.10.6
pydantic-settings==2.7.1

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
pgvector==0.2.4
alembic==1.13.1

# LLM & Embeddings
sentence-transformers==2.3.1
openai==1.58.1
azure-identity==1.19.0
ollama==0.4.5

# Vector Search
faiss-cpu==1.13.2
chromadb==0.4.22

# Data Processing
pandas==2.1.4
numpy==1.26.3
scipy==1.11.4
scikit-learn==1.4.0

# Dashboard
streamlit==1.30.0
plotly==5.18.0

# Task Queue (Optional)
celery==5.3.6
redis==5.0.1

# Utilities
python-dotenv==1.0.0
loguru==0.7.2
requests==2.31.0
httpx==0.27.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Code Quality
black==24.1.1
flake8==7.0.0
isort==5.13.2

# Monitoring & Alerts
python-multipart==0.0.6
aiosmtplib==3.0.1
